moved test into own test files
