+++ floodlight/models/distance_model.py
@@ -139,60 +139,3 @@ class DistanceModel(BaseModel):
     def team_spread(self) -> TeamProperty:
         """Returns the team spread as computed by the fit method."""
         return self._spread
-
-
-##### FOR TESTING
-
-from floodlight.vis.pitches import plot_football_pitch
-from floodlight.vis.positions import plot_positions
-
-def generate_random_xy_data(num_frames, num_points, xlim=(0, 108), ylim=(0, 68)):
-    """Generate random XY data for testing."""
-    data = np.empty((num_frames, num_points * 2))  
-    data[:] = np.nan  # Fill with NaNs
-
-    for frame in range(num_frames):
-        x_coords = np.random.uniform(xlim[0], xlim[1], num_points)
-        y_coords = np.random.uniform(ylim[0], ylim[1], num_points)
-        
-        data[frame, :num_points] = x_coords  # X coordinates for the current frame
-        data[frame, num_points:] = y_coords  # Y coordinates for the current frame
-
-    return XY(data)
-
-def test_distance_model():
-    """Test the DistanceModel class with realistic scenarios including 11 players per team."""
-    
-    # Generate sample data for two teams with 11 players each and 3 frames
-    xy1 = generate_random_xy_data(num_frames=3, num_points=11)
-    xy2 = generate_random_xy_data(num_frames=3, num_points=11, xlim=(50, 158), ylim=(50, 158))
-    
-    # Initialize DistanceModel
-    dm = DistanceModel()
-
-    # Test distance to nearest mate
-    print("Testing distance to nearest mate:")
-    dm.fit(xy1)
-    dtnm = dm.distance_to_nearest_mate()
-    print("Distance to Nearest Mate:")
-    print(dtnm.property)
-
-    # Test team spread
-    print("\nTesting team spread:")
-    spread = dm.team_spread()
-    print("Team Spread:")
-    print(spread.property)
-
-    # Test distance to nearest opponents
-    print("\nTesting distance to nearest opponents:")
-    dm.fit(xy1, xy2)
-    dtno1, dtno2 = dm.distance_to_nearest_opponents()
-    print("Distance to Nearest Opponents (xy1 to xy2):")
-    print(dtno1.property)
-    print("Distance to Nearest Opponents (xy2 to xy1):")
-    print(dtno2.property)
-
-if __name__ == "__main__":
-
-    # Run the distance model test function
-    test_distance_model()
